test_content_summary.py
```python
# Importing necessary modules and functions for testing
import unittest
from unittest.mock import patch, MagicMock
from content_summary import ContentSummary

class TestContentSummary(unittest.TestCase):

    @patch('content_summary.summarize')
    def test_generate_summary(self, mock_summarize):
        # Mocking the summarize function
        mock_summarize.return_value = "This is a summary"

        # Creating a sample content
        content = [{'summary': 'This is a long text that needs to be summarized'}]

        # Creating an instance of ContentSummary
        content_summary = ContentSummary(content)

        # Call the generate_summary method
        summarized_content = content_summary.generate_summary()

        # Assert if the summarize function was called
        mock_summarize.assert_called_once_with('This is a long text that needs to be summarized')

        # Assert if the content was summarized correctly
        self.assertEqual(summarized_content[0]['summary'], "This is a summary")

    def test_summarize_content(self):
        # Creating a sample text
        text = 'This is a short text'

        # Creating an instance of ContentSummary
        content_summary = ContentSummary([])

        # Call the summarize_content method
        summarized_text = content_summary.summarize_content(text)

        # Assert if the text was returned as it is (since it's too short to be summarized)
        self.assertEqual(summarized_text, text)

if __name__ == "__main__":
    unittest.main()
```
